content/xslt/tests/mochitest/test_bug667315.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=667315
-->
<head>
  <title>Test for Bug 667315</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=667315">Mozilla Bug 667315</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 667315 **/

var style =
  '<xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform" ' +
                  'version="1.0">' +
    '<xsl:variable name="var">' +
      '<html><p>a</p></html>' +
    '</xsl:variable>' +
    '<xsl:template match="/">' +
      '<xsl:copy-of select="$var" />' +
    '</xsl:template>' +
  '</xsl:stylesheet>';
var styleDoc = new DOMParser().parseFromString (style, "text/xml");

var data = '<root/>';
var originalDoc = new DOMParser().parseFromString(data, "text/xml");

var processor = new XSLTProcessor();
processor.importStylesheet(styleDoc);

var doc = processor.transformToDocument(originalDoc);
ok(doc instanceof HTMLDocument, "should have switched to html output method");

</script>
</pre>
</body>
</html>
